adding calculus task 9 fixed

+++ math/calculus/9-sum_total.py
@@ -8,4 +8,4 @@ def summation_i_squared(n, sum=0):
     if type(n) is not int:
         return None
     else:
-        return int(n**3/3 + n**2/2 + n/6)
+        return n**3/3 + n**2/2 + n/6
